requirements.txt
aiogram==3.13.1
pytesseract==0.3.13
Pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==2.1.3
python-dotenv==1.0.1
